Rename GlyphPathUtil to ExporterUtils and deduplicate shared SVG and PDF helper functions.
