+++ src/pagx/utils/ExporterUtils.cpp
@@ -16,7 +16,8 @@
 //
 /////////////////////////////////////////////////////////////////////////////////////////////////
 
-#include "GlyphPathUtil.h"
+#include "pagx/utils/ExporterUtils.h"
+#include <cmath>
 #include "base/utils/MathUtil.h"
 #include "pagx/nodes/Font.h"
 #include "pagx/nodes/GlyphRun.h"
@@ -24,6 +25,70 @@
 #include "pagx/nodes/Text.h"
 
 namespace pagx {
+
+using pag::DegreesToRadians;
+using pag::FloatNearlyZero;
+
+FillStrokeInfo CollectFillStroke(const std::vector<Element*>& contents) {
+  FillStrokeInfo info = {};
+  for (const auto* element : contents) {
+    if (element->nodeType() == NodeType::Fill && !info.fill) {
+      info.fill = static_cast<const Fill*>(element);
+    } else if (element->nodeType() == NodeType::Stroke && !info.stroke) {
+      info.stroke = static_cast<const Stroke*>(element);
+    } else if (element->nodeType() == NodeType::TextBox && !info.textBox) {
+      info.textBox = static_cast<const TextBox*>(element);
+    }
+    if (info.fill && info.stroke && info.textBox) {
+      break;
+    }
+  }
+  return info;
+}
+
+Matrix BuildLayerMatrix(const Layer* layer) {
+  Matrix m = layer->matrix;
+  if (layer->x != 0.0f || layer->y != 0.0f) {
+    m = Matrix::Translate(layer->x, layer->y) * m;
+  }
+  return m;
+}
+
+Matrix BuildGroupMatrix(const Group* group) {
+  bool hasAnchor = !FloatNearlyZero(group->anchor.x) || !FloatNearlyZero(group->anchor.y);
+  bool hasPosition = !FloatNearlyZero(group->position.x) || !FloatNearlyZero(group->position.y);
+  bool hasRotation = !FloatNearlyZero(group->rotation);
+  bool hasScale =
+      !FloatNearlyZero(group->scale.x - 1.0f) || !FloatNearlyZero(group->scale.y - 1.0f);
+  bool hasSkew = !FloatNearlyZero(group->skew);
+
+  if (!hasAnchor && !hasPosition && !hasRotation && !hasScale && !hasSkew) {
+    return {};
+  }
+
+  Matrix m = {};
+  if (hasAnchor) {
+    m = Matrix::Translate(-group->anchor.x, -group->anchor.y);
+  }
+  if (hasScale) {
+    m = Matrix::Scale(group->scale.x, group->scale.y) * m;
+  }
+  if (hasSkew) {
+    m = Matrix::Rotate(group->skewAxis) * m;
+    Matrix shear = {};
+    shear.c = std::tan(DegreesToRadians(group->skew));
+    m = shear * m;
+    m = Matrix::Rotate(-group->skewAxis) * m;
+  }
+  if (hasRotation) {
+    m = Matrix::Rotate(group->rotation) * m;
+  }
+  if (hasPosition) {
+    m = Matrix::Translate(group->position.x, group->position.y) * m;
+  }
+
+  return m;
+}
 
 std::vector<GlyphPath> ComputeGlyphPaths(const Text& text, float textPosX, float textPosY) {
   std::vector<GlyphPath> result;
